services/market-data/requirements.txt
# Market Data Service Dependencies

# Data fetching
requests==2.31.0
polygon-api-client==1.13.2
setuptools>=65.0.0  # Required by polygon-api-client

# Data processing
pandas==2.1.4
numpy==1.26.2
ta==0.11.0  # Technical analysis library

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
